+++ concordium-sdk/src/main/java/com/concordium/sdk/requests/ConsensusDetailedStatusQuery.java
@@ -1,0 +1,17 @@
+package com.concordium.sdk.requests;
+
+import com.concordium.sdk.responses.GenesisIndex;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ConsensusDetailedStatusQuery {
+
+    /**
+     * If specified, this determines the genesis index to get status for.
+     * If not specified, the status is returned for the latest genesis index.
+     */
+    @Builder.Default
+    private final GenesisIndex genesisIndex = null;
+}