+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/Baker.java
@@ -2,15 +2,10 @@ package com.concordium.sdk.responses.accountinfo;
 
 import com.concordium.sdk.crypto.bls.BLSPublicKey;
 import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
-import com.concordium.sdk.crypto.elgamal.ElgamalPublicKey;
-import com.concordium.sdk.crypto.pointchevalsanders.PSPublicKey;
 import com.concordium.sdk.crypto.vrf.VRFPublicKey;
 import com.concordium.sdk.responses.BakerId;
 import com.concordium.sdk.responses.bakersrewardperiod.BakerInfo;
 import com.concordium.sdk.transactions.CCDAmount;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -19,6 +14,7 @@ import lombok.extern.jackson.Jacksonized;
 
 import java.util.Optional;
 
+@Getter
 @ToString
 @EqualsAndHashCode
 @Jacksonized
@@ -28,33 +24,37 @@ public final class Baker {
     /**
      * The staked amount.
      */
-    @Getter
     private final CCDAmount stakedAmount;
 
     /**
      * Whether earnings should be restaked automatically or not.
      */
-    @Getter
     private final boolean restakeEarnings;
 
     /**
      * Information about the baker that is staking.
      */
-    @Getter
     private final BakerInfo bakerInfo;
 
     /**
      * The pending changes for the baker.
      * Only present if there is a pending change for the baker.
      */
-    @Getter
     private final Optional<PendingChange> pendingChange;
 
     /**
      * The baker pool info
      */
-    @Getter
     private final BakerPoolInfo bakerPoolInfo;
+
+    /**
+     * A flag indicating whether the account is currently suspended or not. The
+     * flag has a meaning from protocol version 8 onwards. In protocol version 8
+     * it signals whether an account has been suspended and is not participating
+     * in the consensus algorithm. For protocol version < 8 the flag will always
+     * be set to false.
+     */
+    private final boolean isSuspended;
 
     public BakerId getBakerId() {
         return bakerInfo.getBakerId();