+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/updates/queues/PendingUpdateType.java
@@ -23,5 +23,7 @@ public enum PendingUpdateType {
     TimeoutParameters,
     BlockEnergyLimit,
     MinBlockTime,
-    RootKeys
+    RootKeys,
+    FinalizationCommitteeParameters,
+    ValidatorScoreParameters,
 }